+++ measures/sexism-state-paid-family-medical-leave/R/paid_family_leave_proc.R
@@ -49,6 +49,13 @@ state_years <- state_years |>
   mutate(indicator_paid_fam_leave = replace_na(indicator_paid_fam_leave, 0)) |> 
   select(year, statefips, state, indicator_paid_fam_leave)
 
-## ---- Write out the data frame to CSV  ---- 
-state_years |> 
-  write_csv(here::here("measures", topic, "data", "output", glue::glue("{topic}.csv")), na='')
+## ---- Write out data frames to CSV ----------
+if(!dir.exists(here::here("measures", topic, "data", "output"))){
+  dir.create(here::here("measures", topic, "data", "output"))
+  
+  state_years |> 
+    write_csv(here::here("measures", topic, "data", "output", glue::glue("{topic}.csv")), na='')
+} else{
+  state_years |> 
+    write_csv(here::here("measures", topic, "data", "output", glue::glue("{topic}.csv")), na='')
+}
